requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
pydub
gtts
